parity

+++ src/main/java/com/revature/pokemondb/models/dtos/UserDTO.java
@@ -7,14 +7,14 @@ import com.revature.pokemondb.models.User;
  * @author Colby Tang
  */
 public class UserDTO {
-	private int userId; 
+	private Long userId;
 	private String username;
 	private String email; 
 	private String token = "";
 
 	public UserDTO () {}
 
-    public UserDTO(int userId, String username, String email) {
+    public UserDTO(Long userId, String username, String email) {
 		this.userId = userId;
 		this.username = username;
 		this.email = email;
@@ -39,14 +39,14 @@ public class UserDTO {
 	/** 
 	 * @return int
 	 */
-	public int getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 	
 	/** 
 	 * @param customer_id
 	 */
-	public void setUserId(int customer_id) {
+	public void setUserId(Long customer_id) {
 		this.userId = customer_id;
 	}
